reorganize the site builder to load attack/framework information in separate files
